2.6.0 release

+++ play.it-2/src/30_icons.sh
@@ -3,6 +3,7 @@
 # NEEDED VARS: PLAYIT_WORKDIR (WRESTOOL_NAME)
 # CALLS: liberror
 extract_icon_from() {
+	[ "$DRY_RUN" = '1' ] && return 0
 	for file in "$@"; do
 		local destination
 		destination="$PLAYIT_WORKDIR/icons"
@@ -35,28 +36,37 @@ extract_icon_from() {
 # USAGE: sort_icons $app[…]
 # NEEDED VARS: APP_ICON_RES (APP_ID) GAME_ID PKG (PKG_PATH)
 sort_icons() {
+	local app
+	local app_id
+	local icon_res
+	local path_icon
+	local pkg_path
 	for app in "$@"; do
 		testvar "$app" 'APP' || liberror 'app' 'sort_icons'
 
-		local app_id
 		if [ -n "$(eval printf -- '%b' \"\$${app}_ID\")" ]; then
 			app_id="$(eval printf -- '%b' \"\$${app}_ID\")"
 		else
 			app_id="$GAME_ID"
 		fi
 
-		local icon_res
 		icon_res="$(eval printf -- '%b' \"\$${app}_ICON_RES\")"
-		local pkg_path
 		pkg_path="$(eval printf -- '%b' \"\$${PKG}_PATH\")"
 		[ -n "$pkg_path" ] || missing_pkg_error 'sort_icons' "$PKG"
-		for res in $icon_res; do
-			path_icon="$PATH_ICON_BASE/${res}x${res}/apps"
+		[ "$DRY_RUN" = '1' ] && continue
+		if [ -n "${icon_res##* *}" ]; then
+			path_icon="$PATH_ICON_BASE/${icon_res}x${icon_res}/apps"
 			mkdir --parents "${pkg_path}${path_icon}"
-			for file in "$PLAYIT_WORKDIR"/icons/*${res}x${res}x*.png; do
-				mv "$file" "${pkg_path}${path_icon}/${app_id}.png"
+			mv "$file" "${pkg_path}${path_icon}/${app_id}.png"
+		else
+			for res in $icon_res; do
+				path_icon="$PATH_ICON_BASE/${res}x${res}/apps"
+				mkdir --parents "${pkg_path}${path_icon}"
+				for file in "$PLAYIT_WORKDIR"/icons/*${res}x${res}x*.png; do
+					mv "$file" "${pkg_path}${path_icon}/${app_id}.png"
+				done
 			done
-		done
+		fi
 	done
 }
 
@@ -69,6 +79,7 @@ extract_and_sort_icons_from() {
 	local pkg_path
 	pkg_path="$(eval printf -- '%b' \"\$${PKG}_PATH\")"
 	[ -n "$pkg_path" ] || missing_pkg_error 'extract_and_sort_icons_from' "$PKG"
+	[ "$DRY_RUN" = '1' ] && return 0
 	for app in "$@"; do
 		testvar "$app" 'APP' || liberror 'app' 'sort_icons'
 		use_archive_specific_value "${app}_ICON"
@@ -101,6 +112,7 @@ move_icons_to() {
 	local destination_path
 	destination_path="$(eval printf -- '%b' \"\$${1}_PATH\")"
 	[ -n "$destination_path" ] || missing_pkg_error 'move_icons_to' "$1"
+	[ "$DRY_RUN" = '1' ] && return 0
 	(
 		cd "$source_path"
 		cp --link --parents --recursive --no-dereference --preserve=links "./$PATH_ICON_BASE" "$destination_path"
@@ -113,6 +125,7 @@ move_icons_to() {
 # USAGE: postinst_icons_linking $app[…]
 # NEEDED VARS: APP_ICONS_LIST APP_ID|GAME_ID APP_ICON APP_ICON_RES PATH_GAME
 postinst_icons_linking() {
+	[ "$DRY_RUN" = '1' ] && return 0
 	for app in "$@"; do
 		# get icons list associated with current application
 		local app_icons_list
@@ -163,6 +176,7 @@ get_icon_from_temp_dir() {
 	local pkg_path
 	pkg_path="$(eval printf -- '%b' \"\$${PKG}_PATH\")"
 	[ -n "$pkg_path" ] || missing_pkg_error 'get_icon_from_temp_dir' "$PKG"
+	[ "$DRY_RUN" = '1' ] && return 0
 	for app in "$@"; do
 		testvar "$app" 'APP' || liberror 'app' 'get_icon_from_temp_dir'
 		unset app_icon_name
